+++ packages/gallery/src/components/gallery/presets/slideshowGallery.js
@@ -14,6 +14,7 @@ export const fixedStyles = {
   scrollDirection: SCROLL_DIRECTION.HORIZONTAL,
   galleryMargin: 0,
   isVertical: false,
+  groupSize: 1,
   groupTypes: '1',
   
   //this params were moved from the presets in layoutHelper and were not tested and checked yet.
@@ -21,7 +22,6 @@ export const fixedStyles = {
   smartCrop: false,
   gallerySize: 550,
   galleryType: 'Strips',
-  groupSize: 1,
   fixedColumns: 1,
   hasThumbnails: false,
   enableScroll: true,