+++ packages/gallery/src/components/item/itemView.js
@@ -545,6 +545,7 @@ class ItemView extends GalleryComponent {
         key={'video' + this.props.idx}
         hover={itemHover}
         imageDimensions={imageDimensions}
+        hasLink={this.itemHasLink()}
         loadingStatus={{
           failed: this.state.failed,
           loaded: this.state.loaded,
@@ -803,10 +804,16 @@ class ItemView extends GalleryComponent {
     );
   }
 
-  getItemContainerStyles() {
-    const { styleParams, linkData, linkUrl } = this.props;
-    const containerStyleByStyleParams = getContainerStyle(styleParams);
+  itemHasLink(){
+    const { linkData, linkUrl } = this.props;
     const itemDoesntHaveLink = linkData.type === undefined && (linkUrl === undefined || linkUrl === ''); //when itemClick is 'link' but no link was added to this specific item
+    return !itemDoesntHaveLink;
+  }
+
+  getItemContainerStyles() {
+    const { styleParams } = this.props;
+    const containerStyleByStyleParams = getContainerStyle(styleParams);
+    const itemDoesntHaveLink = !this.itemHasLink(); //when itemClick is 'link' but no link was added to this specific item
     const itemStyles = {
       overflowY: styleParams.isSlideshow ? 'visible' : 'hidden',
       position: 'absolute',
@@ -1085,8 +1092,10 @@ class ItemView extends GalleryComponent {
   composeItem() {
     const { photoId, id, hash, idx, styleParams, type, url } = this.props;
 
-    //if there is no url for videos and images, we will not render the itemWrapper (but will render the info element if exists, with the whole size of the item)
-    this.hasRequiredMediaUrl = !(!url && (type === 'image' || type === 'picture' || type === 'video'));
+    //if (there is an url for video items and image items) OR text item (text item do not use media url)
+    this.hasRequiredMediaUrl = url || type === 'text';
+    //if titlePlacement !== SHOW_ON_HOVER and !this.hasRequiredMediaUrl, we will NOT render the itemWrapper (but will render the info element with the whole size of the item)
+    const isItemWrapperEmpty = styleParams.titlePlacement !== PLACEMENTS.SHOW_ON_HOVER && !this.hasRequiredMediaUrl;
     const innerDiv = (
       <div
         className={this.getItemContainerClass()}
@@ -1117,7 +1126,7 @@ class ItemView extends GalleryComponent {
             ...(this.props.styleParams.titlePlacement === PLACEMENTS.SHOW_ON_THE_LEFT && {float: 'right'})
           }}
         >
-          {this.hasRequiredMediaUrl && (<div
+          {!isItemWrapperEmpty && (<div
             data-hook="item-wrapper"
             className={this.getItemWrapperClass()}
             key={'item-wrapper-' + id}