+++ packages/gallery/src/settings/index.js
@@ -30,6 +30,9 @@ import textBoxBorderColor from './options/textBoxBorderColor';
 import textBoxBorderRadius from './options/textBoxBorderRadius';
 import textBoxBorderWidth from './options/textBoxBorderWidth';
 import textBoxWidth from './options/textBoxWidth';
+import imageInfoType from './options/imageInfoType';
+import groupSize from './options/groupSize';
+import collageDensity from './options/collageDensity';
 import groupTypes from './options/groupTypes';
 
 export default {
@@ -65,5 +68,8 @@ export default {
     textBoxBorderRadius,
     textBoxBorderWidth,
     textBoxWidth,
+    imageInfoType,
+    groupSize,
+    collageDensity,
     groupTypes,
 }